requirements.txt
PyQt6
google-generativeai
python-dotenv
nltk
